+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+
+# Upcoming
+
+- Add TopNav component
+
+
 # 7.1.1
 
 ## Updates